requirements.txt
Pillow==9.1.1
streamlit==1.9.2
scikit-learn==1.1.1
